+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+0.7.0 / 2018-04-20
+==================
+
+  * Assert.request module
+  * Assert.visible now checks every parent node is visible
+  * Browser.setViewport
+  * Assert.attribute and assert.not.attribute now match any element
+  * Assert.href and assert.not.href now match any element
+  * Assert.visible message is now different when the element doesn't exist
+  * WaitFor message changed when waiting for a function
+
 0.6.0 / 2018-04-13
 ==================